Rename to SamplingFunction
